Initial model testing
Loads the dataset, and runs it through a couple of models for some
preliminary accuracy testing.  There's a little cleanup here where we're
ditching rows with NaNs in the band data.  Looks like where there's one
NaN it's pretty much NaNs all the way through so I'm just scrubbing any
row where there's one.  Much tuning of the models is still
required to make any proper analysis of which is best suited to the
domain. My hunch is that an ANN is probably where we'll want to go with
the high dimensionality and lack of volume in the training data but I'm
no expert here.

Something's definitely up with the SVC approach here though, it's
classifying with all 0 classes, which makes the think that the one-hot
we're using for input vals may not be appropriate, though from what I
recall straight-up integer classes on an SVC ought to be okay.

Random forest appears to be performing pretty well  I'm not sure what
depth is enabled by default with this model but that param is probably
worth toying with.