[WIP] Playing with histogram visualization of our dataset
We're having trouble getting much accuracy out of our models here, so
let's try some visual analysis to see if we can find ways of optimizing
our data set our or approach to modeling.

Here, we introduce a new Visualizer class and invoke it from main, so
running the main.py script by itself generates new visualizations in
/tmp.

This mechanism needs a nicer interface to be suitable for re-use, but
it's sufficient for generating quick plots currently.